backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pyjwt==2.8.0
python-dotenv==1.0.0
bcrypt==4.0.1 